Relax INT_MAX buffer size limit for internode (#229)
* relaxed the buffer size up to INT_MAX limit for internode

* tested over INT_MAX buffer size in exp/grpcoll tests

* minor fixed

* added docstring for config funcs

* added minimium num bytes check for native grpcoll

* fixed tma bytes and num warps for internode cache notify kernel

* raised up default num_rdma_bytes

* further fixed internode cache notify kernel for group reduce

* removed the temp debug code to make benchmark mask split-aligned

+++ magi_attention/csrc/comm/grpcoll/buffer.hpp
@@ -76,12 +76,12 @@ struct Buffer {
   bool low_latency_mode = false;
 
   // NVLink Buffer
-  int64_t num_nvl_bytes;
+  size_t num_nvl_bytes;
   void* buffer_ptrs[NUM_MAX_NVL_PEERS] = {nullptr};
   void** buffer_ptrs_gpu = nullptr;
 
   // NVSHMEM Buffer
-  int64_t num_rdma_bytes;
+  size_t num_rdma_bytes;
   void* rdma_buffer_ptr = nullptr;
 
   // Device info and communication
@@ -120,7 +120,7 @@ struct Buffer {
   int* grpcoll_recv_rdma_counter_mapped = nullptr;
 
  public:
-  Buffer(int rank, int num_ranks, int64_t num_nvl_bytes, int64_t num_rdma_bytes, bool low_latency_mode, bool explicitly_destroy);
+  Buffer(int rank, int num_ranks, size_t num_nvl_bytes, size_t num_rdma_bytes, bool low_latency_mode, bool explicitly_destroy);
 
   ~Buffer() noexcept(false);
